Move out integration test services
